add copyright notices
