+++ website/components/bounties.tsx
@@ -1,7 +1,10 @@
+// Copyright (c) 2023 Algora PBC.
+// https://github.com/algora-io/sdk/blob/5bcda5de293246b959b5e5ea81529ec46d7a8fe3/README.md
+
 import { useEffect, useState } from 'react';
 import { algora, type AlgoraOutput } from '@algora/sdk';
 import Link from 'next/link';
-import { motion, useMotionTemplate, useSpring } from 'framer-motion';
+import { Spotlight } from './spotlight';
 
 const org = 'aidenybai';
 const limit = 2;
@@ -51,50 +54,25 @@ export function Bounties() {
 }
 
 function BountyCard(props: { bounty: Bounty }) {
-  const mouseX = useSpring(0, { stiffness: 500, damping: 100 });
-  const mouseY = useSpring(0, { stiffness: 500, damping: 100 });
-
-  const onMouseMove: React.MouseEventHandler<HTMLElement> = ({
-    currentTarget,
-    clientX,
-    clientY,
-  }) => {
-    const { left, top } = currentTarget.getBoundingClientRect();
-    mouseX.set(clientX - left);
-    mouseY.set(clientY - top);
-  };
-  const maskImage = useMotionTemplate`radial-gradient(240px at ${mouseX}px ${mouseY}px, white, transparent)`;
-  const style = { maskImage, WebkitMaskImage: maskImage };
-
   return (
     <Link
       href={props.bounty.task.url}
       rel="noopener noreferrer"
-      onMouseMove={onMouseMove}
       className="block group relative h-full rounded-lg border border-zinc-400/50 dark:border-purple-500/50 bg-gradient-to-br from-zinc-300/30 via-zinc-300/40 to-zinc-300/50 dark:from-purple-600/20 dark:via-purple-600/30 dark:to-purple-600/40 md:gap-8 transition-colors hover:border-zinc-400 hover:dark:border-purple-500 hover:bg-zinc-300/10 hover:dark:bg-zinc-600/5 !no-underline"
     >
-      <div className="pointer-events-none">
-        <div className="absolute inset-0 z-0 opacity-0 transition-opacity [mask-image:linear-gradient(black,transparent)] group-hover:opacity-100" />
-        <motion.div
-          className="absolute inset-0 z-10 bg-zinc-300/30 dark:bg-purple-600/20 opacity-0 transition-opacity group-hover:opacity-100"
-          style={style}
-        />
-        <motion.div
-          className="absolute inset-0 z-10 opacity-0 mix-blend-overlay transition-opacity group-hover:opacity-100"
-          style={style}
-        />
-      </div>
-      <div className="relative h-full p-4">
-        <div className="text-2xl font-bold text-green-500 group-hover:text-green-600 dark:text-green-400 dark:group-hover:text-green-300 transition-colors">
-          {props.bounty.reward_formatted}
+      <Spotlight className="bg-zinc-300/30 dark:bg-purple-600/20">
+        <div className="h-full p-4">
+          <div className="text-2xl font-bold text-green-500 group-hover:text-green-600 dark:text-green-400 dark:group-hover:text-green-300 transition-colors">
+            {props.bounty.reward_formatted}
+          </div>
+          <div className="mt-0.5 text-sm text-zinc-700 dark:text-purple-200 group-hover:text-zinc-800 dark:group-hover:text-purple-100 transition-colors">
+            {props.bounty.task.repo_name}#{props.bounty.task.number}
+          </div>
+          <div className="mt-3 line-clamp-1 break-words text-lg font-medium leading-tight text-zinc-800 dark:text-purple-50 group-hover:text-zinc-900 dark:group-hover:text-white">
+            {props.bounty.task.title}
+          </div>
         </div>
-        <div className="mt-0.5 text-sm text-zinc-700 dark:text-purple-200 group-hover:text-zinc-800 dark:group-hover:text-purple-100 transition-colors">
-          {props.bounty.task.repo_name}#{props.bounty.task.number}
-        </div>
-        <div className="mt-3 line-clamp-1 break-words text-lg font-medium leading-tight text-zinc-800 dark:text-purple-50 group-hover:text-zinc-900 dark:group-hover:text-white">
-          {props.bounty.task.title}
-        </div>
-      </div>
+      </Spotlight>
     </Link>
   );
 }
